+++ tests/testthat/test-oauth-flow-auth-code.R
@@ -22,7 +22,8 @@ test_that("so-called 'hosted' sessions are detected correctly", {
 
 test_that("URL embedding authorisation code and state can be input manually", {
   local_mocked_bindings(
-    readline = function(prompt = "") "https://x.com?code=code&state=state"
+    askpass = function(prompt = "") "https://x.com?code=code&state=state",
+    .package = "askpass"
   )
   expect_equal(oauth_flow_auth_code_read("state"), "code")
   expect_error(oauth_flow_auth_code_read("invalid"), "state does not match")
@@ -32,7 +33,8 @@ test_that("JSON-encoded authorisation codes can be input manually", {
   input <- list(state = "state", code = "code")
   encoded <- openssl::base64_encode(jsonlite::toJSON(input))
   local_mocked_bindings(
-    readline = function(prompt = "") encoded
+    askpass = function(prompt = "") encoded,
+    .package = "askpass"
   )
   expect_equal(oauth_flow_auth_code_read("state"), "code")
   expect_error(oauth_flow_auth_code_read("invalid"), "state does not match")
@@ -42,14 +44,15 @@ test_that("bare authorisation codes can be input manually", {
   state <- base64_url_rand(32)
   sent_code <- FALSE
   local_mocked_bindings(
-    readline = function(prompt = "") {
+    askpass = function(prompt = "") {
       if (sent_code) {
         state
       } else {
         sent_code <<- TRUE
         "zyx987"
       }
-    }
+    },
+    .package = "askpass"
   )
   expect_equal(oauth_flow_auth_code_read(state), "zyx987")
   expect_error(oauth_flow_auth_code_read("invalid"), "state does not match")